+++ src/github-usernames-link-provider.ts
@@ -1,0 +1,55 @@
+import vscode from "vscode"
+
+function githubUserToUrl(username: string): vscode.Uri {
+  const isTeamName = username.includes("/")
+
+  if (isTeamName) {
+    const [org, name] = username.split(/\//)
+    return vscode.Uri.parse(`https://github.com/orgs/${org}/teams/${name}`)
+  }
+  return vscode.Uri.parse(`https://github.com/${username}`)
+}
+
+/**
+ * Add links to usernames in CODEOWNERS file that open on GitHub.
+ */
+export class GitHubUsernamesLinkProvider
+  implements vscode.DocumentLinkProvider
+{
+  provideDocumentLinks(
+    document: vscode.TextDocument,
+  ): vscode.ProviderResult<vscode.DocumentLink[]> {
+    const regex = new RegExp(/\S*@\S+/g)
+    const text = document.getText()
+    let matches: RegExpExecArray | null = null
+    const links = []
+    // loop copied from https://github.com/microsoft/vscode-extension-samples/blob/dfb20f12d425bad2ede0f1faae25e0775ca750eb/codelens-sample/src/CodelensProvider.ts#L24-L37
+    while ((matches = regex.exec(text)) !== null) {
+      const line = document.lineAt(document.positionAt(matches.index).line)
+      const indexOf = line.text.indexOf(matches[0])
+      const position = new vscode.Position(line.lineNumber, indexOf)
+      const range = document.getWordRangeAtPosition(
+        position,
+        new RegExp(/\S*@\S+/g),
+      )
+
+      if (range) {
+        const username = document.getText(range)
+
+        // don't make emails clickable
+        // e.g. docs@example.com
+        if (!username.startsWith("@")) {
+          continue
+        }
+        const link = new vscode.DocumentLink(
+          range,
+          githubUserToUrl(username.replace(/^@/, "")),
+        )
+        link.tooltip = `View ${username} on Github`
+
+        links.push(link)
+      }
+    }
+    return links
+  }
+}
